Solution divided in 2 projects: GUI & Core
