+++ .gitignore
@@ -13,3 +13,4 @@
 /XstReader.Core/bin
 /XstReader.Core/obj
 /XstReader.csproj.user
+.vs/XstReader/v16/.suo
